# feat: lazy initialization codec