--- a/internal/encoding/toml/codec.go
+++ b/internal/encoding/toml/codec.go
@@ -2,15 +2,20 @@ package toml
 
 import (
 	"github.com/pelletier/go-toml/v2"
+	"github.com/spf13/viper/internal/encoding/codec"
 )
 
-// Codec implements the encoding.Encoder and encoding.Decoder interfaces for TOML encoding.
+// Codec implements the encoding.Codec interface for TOML encoding.
 type Codec struct{}
 
-func (Codec) Encode(v map[string]interface{}) ([]byte, error) {
+func New(_ ...interface{}) codec.Codec {
+	return &Codec{}
+}
+
+func (*Codec) Encode(v map[string]interface{}) ([]byte, error) {
 	return toml.Marshal(v)
 }
 
-func (Codec) Decode(b []byte, v map[string]interface{}) error {
+func (*Codec) Decode(b []byte, v map[string]interface{}) error {
 	return toml.Unmarshal(b, &v)
 }