… internal/encoding/decoder.go
… internal/encoding/decoder_test.go
… internal/encoding/encoder.go
… internal/encoding/encoder_test.go