@xieyuschen 2023-08-18
> Wow, thanks @xieyuschen !
> 
> Give me some time to review this PR, but I really appreciate the contribution.
> 
> One note off the bat: performance generally is not a huge concern in case of Viper as it's mostly initialized/used at the beginning of an application run, so readability is generally preferred over rigorous performance optimization.

I agree with you, and I have changed the implementation in a go style. @sagikazarmark Could you kindly check it if you have time?

Regards